+++ test/index-test.js
@@ -1,0 +1,13 @@
+const chai = require('chai');
+const assert = chai.assert;
+
+const PlayGolf = require('../lib/play-golf');
+const Menu = require('../lib/menu');
+
+describe("GamePlay", function() {
+
+});
+
+describe("Game Menu", function() {
+
+});
